Add files via upload
This lab focused on Assessing Network with Common Security Tools. The time needed to completed was unlimited. The work just had to be submitted before due date
so I took my time not only to complete the lab for submission but to also learn and research while I perform the lab.
Some security and network tools used were Linux Terminal, Windows CMD, Wireshark and Zenmap. I also used a spread to document findings during the assessment.