Add files via upload
This lab demonstrate Configuring a VPN Client for Secure File Transfer. We are living in moment in time where the internet has made almost everything virtually possible.
This makes communication and the internet all together very convenient and efficient but critical and expose to security threat. For this reason, It is very important that Files shared internally and externally are secured. 
In this Lab, I configured a VPN client for a remote user using IPSEC.
I was also able to capture all secure packets traveling through the network as shown in PAGE 4. 
For detailed information, Please follow the results in the pdf file.