Add files via upload
This is a demonstration on configuring an IPsec VPN Server. VPNs are very necessary especially when we want to allow remote connection to a private network.
A remote access or public access to a private network poses so many threats to the internal network and so it is very important to configure a secure VPN connection so that packets can be encrypted. This is very important especially if a data in transit is captured by a third party or a threat agent. 
In this lab, I created a Certificate Authority to issue certificates.
I also created various VPN firewall rules like Allowing connections to the VPN server process from clients anywhere on the internet. I also allowed all traffic from connected clients to pass inside the VPN tunnel. This means, as long as you are connected to the VPN server, all traffics will pass through the secure tunnel and make sure data is encrypted!