Add files via upload
Human Beings are the weakest link in the security chain. With that being said, No matter the number of security appliances we put in place to protect a network, data and information, End user training must also be as important .This lab demonstrate how tools like LinkedIn and target website can be used to perform reconnaissance in order to gather more information to conduct a social engineering attack.