Add files via upload
Penetration Test is one way of assessing a network to find weak areas and vulnerabilities so that it can be mitigated. Security is everyday practice and so conducting regular penetration test is highly recommended. In this lab, I conducted a penetration test on pfsense firewall with tools like Nessus,Nmap and OpenVAS. During the penetration testing, I discovered the DMZ posses some vulnerabilities. For instance the Jquery hosted on the remote server needed an update. Also a. better SSL Certificate is needed on the Secure lab